Add host builder extensions for Omny builder package
